taskmanager/requirements.txt
Django>=3.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-filter
django-cors-headers
drf-yasg
